requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
billiard==4.2.3
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.5
idna==3.11
inflection==0.5.1
jiter==0.12.0
kombu==5.5.4
numpy==2.3.5
openai==2.8.1
packaging==25.0
pandas==2.3.3
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14





















# Step 2 — Build Data Ingestion Module

# This is the heart of the project.

# Goal:

# User uploads:

# CSV

# Excel

# JSON

# PDF

# Then you store it & prepare for analysis.

# ✔ Tasks:

# Create a model:

# class UploadedFile(models.Model):
#     user = models.ForeignKey(User, on_delete=models.CASCADE)
#     file = models.FileField(upload_to='uploads/')
#     uploaded_at = models.DateTimeField(auto_now_add=True)


# Create Upload API

# Validate file type

# Process files using Pandas

# Return cleaned data summary

# total rows

# total columns

# column types

# missing values

# 👉 This powers the analysis engine.

# ⭐ Step 3 — Build Data Analysis Module

# Once data is uploaded → user triggers analysis.

# ✔ Tasks:

# Accept file ID

# Perform basic EDA:

# summary stats

# missing data

# correlations

# Generate insights using ML/A.I:

# anomaly detection

# trend detection

# segmentation

# Return as structured JSON and visual-ready data